introduce analysed ASM struct, wip
