+++ ast/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ast"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+num-bigint = "0.4.3"
+number = { path = "../number" }
+
