+++ executor/src/constant_evaluator/mod.rs
@@ -1,8 +1,9 @@
 use std::collections::HashMap;
 
+use ast::parsed::{BinaryOperator, UnaryOperator};
 use itertools::Itertools;
 use number::{DegreeType, FieldElement};
-use pil_analyzer::{Analyzed, BinaryOperator, Expression, FunctionValueDefinition, UnaryOperator};
+use pil_analyzer::{Analyzed, Expression, FunctionValueDefinition};
 use rayon::prelude::{IntoParallelIterator, ParallelIterator};
 
 /// Generates the constant polynomial values for all constant polynomials
