+++ pilgen/Cargo.toml
@@ -9,4 +9,8 @@ number = { path = "../number" }
 parser_util = { path = "../parser_util" }
 parser = { path = "../parser" }
 itertools = "0.10.5"
-pretty_assertions = "1.3.0"
+num-bigint = "0.4.3"
+ast = { path = "../ast" }
+
+[dev-dependencies]
+analysis = { path = "../analysis" }
