+++ test_data/asm/batching/disjoint_assignment_reg.asm
@@ -91,23 +91,5 @@ addr <=Y= wrap(tmp1); // Y -> X -> addr
 mstore tmp1; // tmp1 -> X -> ()
 // END BATCH Label
 
-// if we're using inline pil, any column referenced both there and in an expression needs to be added to the footprint (recursively)
-connected_by_inline_pil::
-// C and D are linked in inline PIL
-reg C;
-reg D;
-pil {
- col fixed FIRST = [1] + [0]*;
- FIRST * C = 2;
- FIRST * D = 2;
- C*D = 4;
-}
-// we change C
-C <=X= 1;
-// C*D should break here, which is why we cannot batch with the next
-// we change D
-D <=Y= 4;
-// C*D is verified again here
-
 end::
  loop;
